requirements.txt
torch
torchvision
foolbox
numpy
pillow
streamlit


